Extract statements_llm.go
